app/backend/requirements.clean.txt
pydantic-ai
pydantic-ai-slim[openai]
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
chromadb
sentence-transformers
langchain
langchain-openai
langchain-community
langchain-huggingface
pdf2image
python-multipart
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pydantic
pydantic-settings
